pyproject.toml
[project]
name = "amnezia-api"
version = "0.1.0"
description = ""
authors = [
    {name = "homostultus39",email = "homostultus39@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "docker (>=7.1.0,<8.0.0)",
    "psutil (>=7.2.2,<8.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
]

[tool.poetry]
packages = [{include = "amnezia_api", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
